requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
pandas==2.0.0
numpy==1.24.0
gunicorn==21.2.0
whitenoise==6.5.0